move the sub edit mode into mode state namespace.
