+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/States/Modes/LanguageEditMode.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Languages
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes
 {
     public enum LanguageEditMode
     {
