+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/States/Modes/NoteEditMode.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.RubyRomaji
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes
 {
-    public enum TextTagEditMode
+    public enum NoteEditMode
     {
         Generate,
