+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/States/Modes/TextTagEditMode.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes
 {
-    public enum NoteEditMode
+    public enum TextTagEditMode
     {
         Generate,
